Django-Middleware-0x03/requirements.txt
# requirements.txt

Django
djangorestframework
django-cors-headers
channels
channels-redis
redis
python-dotenv
djangorestframework-simplejwt
psycopg2-binary --only-binary=:all
Django_Filter
